iclr_retouchllm/requirements.txt
# Core dependencies
numpy>=1.21.0
opencv-python>=4.5.0
Pillow>=9.0.0

# Deep learning
torch>=2.0.0
transformers>=4.40.0

# Vision & metrics
lpips>=0.1.4
scikit-image>=0.19.0
ftfy
regex
# OpenAI CLIP: pip install git+https://github.com/openai/CLIP.git

# API client (OpenAI 格式)
openai>=1.0.0

# Utilities
tqdm>=4.60.0
